Auth Store with Pinia
This is a Pinia store for managing authentication-related state and actions in the Blogy app. It includes functions for authenticating users, setting user data, signing out, and checking for an existing token in the local storage.

Key Features:

- Authentication: The authenticateUser function is used to authenticate users by sending their credentials to the server. It sets the user data using the setUser function when the authentication is successful.

- Setting User Data: The setUser function is responsible for setting the user data after successful authentication. It decodes the JWT token, stores it in the local storage, and sets the user state.

- Sign Out: The signOut function clears the user data by removing the JWT token from the local storage and setting the user state to null. It then navigates the user to the sign-in page.

- Check for Token: The checkForToken computed property is used to check for an existing token in the local storage. If a token is found, it calls the setUser function to set the user data. If no token is found, it calls the signOut function to log the user out.

- Router Integration: The store utilizes the Vue Router (useRouter) to handle navigation after successful authentication or sign-out.

Overall, this Pinia store provides a centralized and well-organized way to manage authentication-related state and actions in the Blogy app. It ensures that the user data is set correctly after authentication, and it handles sign-out and token checks efficiently.